feat(endpoint): move setPresence endpoint to instance
