+++ src/whatsapp/routers/instance.router.ts
@@ -3,9 +3,9 @@ import { RequestHandler, Router } from 'express';
 import { Auth, ConfigService, Database } from '../../config/env.config';
 import { Logger } from '../../config/logger.config';
 import { dbserver } from '../../libs/db.connect';
-import { instanceNameSchema, oldTokenSchema } from '../../validate/validate.schema';
+import {instanceNameSchema, oldTokenSchema, presenceOnlySchema} from '../../validate/validate.schema';
 import { RouterBroker } from '../abstract/abstract.router';
-import { InstanceDto } from '../dto/instance.dto';
+import { InstanceDto, SetPresenceDto } from '../dto/instance.dto';
 import { OldToken } from '../services/auth.service';
 import { instanceController } from '../whatsapp.module';
 import { HttpStatus } from './index.router';
@@ -97,6 +97,22 @@ export class InstanceRouter extends RouterBroker {
         });
 
         return res.status(HttpStatus.OK).json(response);
+      })
+      .post(this.routerPath('setPresence'), ...guards, async (req, res) => {
+        logger.verbose('request received in setPresence');
+        logger.verbose('request body: ');
+        logger.verbose(req.body);
+
+        logger.verbose('request query: ');
+        logger.verbose(req.query);
+        const response = await this.dataValidate<null>({
+          request: req,
+          schema: presenceOnlySchema,
+          ClassRef: SetPresenceDto,
+          execute: (instance, data) => instanceController.setPresence(instance, data),
+        });
+
+        return res.status(HttpStatus.CREATED).json(response);
       })
       .delete(this.routerPath('logout'), ...guards, async (req, res) => {
         logger.verbose('request received in logoutInstances');
